Move JextractToolRunner and JextractApiTestBase in lib folder
