+++ test/java/TEST.properties
@@ -1,0 +1,1 @@
+lib.dirs = /lib
