+++ test/java/org/openjdk/jextract/test/api/SmokeTest.java
@@ -25,6 +25,7 @@ package org.openjdk.jextract.test.api;
 import org.openjdk.jextract.Declaration;
 import org.openjdk.jextract.Type;
 import org.testng.annotations.Test;
+import testlib.JextractApiTestBase;
 
 public class SmokeTest extends JextractApiTestBase {
