+++ test/java/org/openjdk/jextract/test/api/Test8239490.java
@@ -25,6 +25,7 @@ package org.openjdk.jextract.test.api;
 import org.openjdk.jextract.Declaration;
 import org.openjdk.jextract.Type;
 import org.testng.annotations.Test;
+import testlib.JextractApiTestBase;
 
 public class Test8239490 extends JextractApiTestBase {
     @Test
