+++ test/java/org/openjdk/jextract/test/api/TestAttributes.java
@@ -32,6 +32,7 @@ import java.util.stream.Stream;
 import org.openjdk.jextract.Declaration;
 import org.openjdk.jextract.Type;
 import org.testng.annotations.Test;
+import testlib.JextractApiTestBase;
 
 import static org.testng.Assert.assertEquals;
