+++ test/java/org/openjdk/jextract/test/toolprovider/Test7903148.java
@@ -25,17 +25,11 @@ package org.openjdk.jextract.test.toolprovider;
 
 import org.openjdk.jextract.test.TestUtils;
 import org.testng.annotations.Test;
+import testlib.JextractToolRunner;
+
 import java.nio.file.Path;
 import static org.testng.Assert.assertNotNull;
 
-/*
- * @test
- * @library /test/lib
- * @build JextractToolRunner
- * @bug 7903148
- * @summary jextract crashes when C11 keyword _Static_Assert is used in a header
- * @run testng/othervm --enable-native-access=ALL-UNNAMED Test7903148
- */
 public class Test7903148 extends JextractToolRunner {
     @Test
     public void test() {
