+++ test/java/org/openjdk/jextract/test/toolprovider/Test8240811.java
@@ -30,6 +30,8 @@ import jdk.incubator.foreign.GroupLayout;
 import jdk.incubator.foreign.MemoryLayout;
 import org.openjdk.jextract.test.TestUtils;
 import org.testng.annotations.Test;
+import testlib.JextractToolRunner;
+
 import static org.testng.Assert.assertNotNull;
 import static org.testng.Assert.assertTrue;
