+++ test/java/org/openjdk/jextract/test/toolprovider/Test8245767.java
@@ -28,6 +28,8 @@ import java.nio.file.Path;
 
 import org.openjdk.jextract.test.TestUtils;
 import org.testng.annotations.Test;
+import testlib.JextractToolRunner;
+
 import static org.testng.Assert.assertNull;
 import static org.testng.Assert.assertNotNull;
 import static org.testng.Assert.assertTrue;
