+++ test/java/org/openjdk/jextract/test/toolprovider/Test8258405.java
@@ -27,6 +27,8 @@ import java.nio.file.Path;
 
 import org.openjdk.jextract.test.TestUtils;
 import org.testng.annotations.Test;
+import testlib.JextractToolRunner;
+
 import static org.testng.Assert.assertNotNull;
 
 public class Test8258405 extends JextractToolRunner {
