+++ test/java/org/openjdk/jextract/test/toolprovider/Test8260344.java
@@ -27,6 +27,7 @@ import java.nio.file.Path;
 
 import org.openjdk.jextract.test.TestUtils;
 import org.testng.annotations.Test;
+import testlib.JextractToolRunner;
 
 public class Test8260344 extends JextractToolRunner {
     @Test
