+++ test/java/org/openjdk/jextract/test/toolprovider/Test8260705.java
@@ -29,6 +29,8 @@ import jdk.incubator.foreign.MemoryAddress;
 import jdk.incubator.foreign.MemorySegment;
 import org.openjdk.jextract.test.TestUtils;
 import org.testng.annotations.Test;
+import testlib.JextractToolRunner;
+
 import static org.testng.Assert.assertNotNull;
 import static org.testng.Assert.assertTrue;
