+++ test/java/org/openjdk/jextract/test/toolprovider/Test8260717.java
@@ -28,6 +28,7 @@ import jdk.incubator.foreign.MemoryAddress;
 import jdk.incubator.foreign.MemorySegment;
 import org.openjdk.jextract.test.TestUtils;
 import org.testng.annotations.Test;
+import testlib.JextractToolRunner;
 
 public class Test8260717 extends JextractToolRunner {
     @Test
