+++ test/java/org/openjdk/jextract/test/toolprovider/Test8262851.java
@@ -25,6 +25,8 @@ package org.openjdk.jextract.test.toolprovider;
 
 import org.openjdk.jextract.test.TestUtils;
 import org.testng.annotations.Test;
+import testlib.JextractToolRunner;
+
 import java.nio.file.Path;
 import static org.testng.Assert.assertNotNull;
