+++ test/java/org/openjdk/jextract/test/toolprovider/TestAttributedPointerTypedef.java
@@ -23,6 +23,7 @@
 package org.openjdk.jextract.test.toolprovider;
 
 import org.testng.annotations.Test;
+import testlib.JextractToolRunner;
 
 public class TestAttributedPointerTypedef extends JextractToolRunner {
     @Test
