+++ test/java/org/openjdk/jextract/test/toolprovider/UniondeclTest.java
@@ -29,6 +29,8 @@ import org.openjdk.jextract.test.TestUtils;
 import org.testng.annotations.Test;
 import java.nio.file.Path;
 import jdk.incubator.foreign.GroupLayout;
+import testlib.JextractToolRunner;
+
 import static org.testng.Assert.assertNotNull;
 import static org.testng.Assert.assertTrue;
